Remove obsolete and outdated sections from check manpages
Information about the configuration is now found in WATO. If you
need the correct syntax of a check configuration then please
create a configuration example with WATO and look into the <tt>rules.mk</tt>
file as a reference. Also the information about the metrics has been dropped
since the new metric system has a correct description for each metric.