3951 FIX oracle_sessions: fixed bug with default absolute levels
When relative levels are given in a rule but no absolute levels, the default levels from the checkplugin are still used.
This change is compatible and will ignore absolute levels in a given rule, if not explicitly set. Defaults for absolute levels are still used if no rule is given.