4035 FIX df_zos: Fixed ignoring last filesystem after new discovery
Filesystemcheck df_zos was ignoring the last filesystem, because not all versions
of z/OS will end with a zero line after the df command. This was fixed with adding
the pattern line fix at the end of the agent output.

Pleae do a new discovery to check, if there is a filesystem is missing.