Removed useless assignments.
Zeroing parameters or local variables doesn't really make sense, unless we
capture the current environment in a closure and store that away somehow.
This is quite rare, and more often than not, a better solution is to lift
out the closure.

Setting something to null/undefined doesn't even speed up GC, because every
GC implementation worth its salt doesn't to naive reference counting,
anyway, and smears/delays its execution.  So we are left with 2 scenarios:

   * The JIT/interpreter is stupid and actually executes the assignment:
     Then we do useless work and confuse the human reader.

   * The JIT/interpreter does its job: The assignment will be nuked by dead
     code elimination, but we still confuse the human reader.