TextAreaUnicode can now use available width, if Browser supports it
The ValueSpec TextAreaUnicode has the new argument "try_max_width". If it
is set to True, then the size of the area is allowed to grow beyond the
arguments cols (which is now a minimum size). This is done by using calc()
in CSS. Older browsers that do not handle calc() should simply ignore this
and keep to the old behaviour.