4154 Do not send notifications if host found and currently in downtime
When you forward notifications from the Event Console to Check_MK then
from now on these are being suppressed if the host can be found in the
active monitoring and is currently in a scheduled downtime.

This is very much what users except and want to happen. Please note that
- however - this is an incompatible change sind now some notifications
will not longer be sent that have been sent previously.

Change-Id: Iee146930e13e2b382193c6f4c1abcf8e9b7d4808