4155 FIX Do not perform actions on events if overflow limit is active
If the EC blocks a new event because of an active overflow limit
then it now correctly also omit any configured actions (i.e.
a notification) on that event.

Previously actions would be executed anyway - that way undermining
the overflow protection. Furthermore the action "Send notifiction"
would run into an exception because of a missing event ID.

Change-Id: I5f4f3902a336f9c635544027933f0d53e51fb359