4207 FIX mk_jolokia: Use configured port as default instance name
Previously, the mk_jolokia plugin would default to the port as
an instance name if no name was configured before reading the
config file. This led to the default instance name always being
8080, regardless of which port was actually configured.

If you have configured a nonstandard port in jolokia.cfg,
but not an instance name, rediscovery is required, as well as
adjusting rules relying on the instance name.

Change-Id: I0ee9508971449b85ad5adb66c8b436c8b07a635e