5591 Only load graphs that are currently visible
With this change the graphs shown by the GUI are only rendered
when they are (partially) visible to the user. The initial loading
of the graphs out of the viewport is now being delayed until they
become visible to the user.

This is a performance improvement for views showing a lot of graphs.

Change-Id: I087d777ccc99e28f131f4d7c90bd22a690b122c8