4227 New manual WATO configuration lock mechanism
It is now possible to manually lock the WATO configuration for modifications and set
the configuration to read only mode.

You can find the option to configure it below <i>Global Settings > Read only mode</i>.

On that page the WATO configuration can be set to read only mode for all users that are not
permitted to ignore the read only mode. All users that are permitted to set the read only mode
can disable it again when another permitted user enabled it before.

The read only mode can be set permanently or in a specific time window. You can set a message
which is then shown to all users while the read only mode is active.

Change-Id: I94fac7c7c23d49444016dc1ea506cbaa2a61433f