mkbench: Started creating a standard system benchmark tool
The tool is meant to execute different tasks in a Check_MK site
on a system and measure the performance of the task on that
system. The test is saving different static information of
the system (Installed CPUs, memory, disks, ...) and captures
the most important peformance relevant values during testing
for later analysis and comparision.

The tool also monitors the performance of different values of
the system and aborts tests in case thresholds are reached.

Change-Id: I128418d7897ab318ea888029402131defb6f5c15