4017 New option to hide WATO folders when the user has insufficent permissions
The new global settings option <i>Hide folders without read permissions</i> allows
you to hide WATO folders from users. If a user does not have read permission to the
subfolder AND any of its subfolders, the folder is simply not shown.

Note: This feature only reduces the number of folders in the WATO folder overview.
It does not hide these folders in other WATO pages.

Change-Id: Ib483c02b4f5818faac3b8d440fa28ebdf95362f4