4194 FIX esx_vsphere_hostsystem.multipath: Fixed grouping of paths
The changes in vSphere6 made it impossible to display the data in the known way. Now it's necessary to display the information as L#. Additionally there is a suffix to identify the type of the LUN correctly. Otherwise the path to a pseudo-logical device like USB sticks could be accidentally added to the paths of a SAN-LUN.

A rediscovery will be necessary since this fix will change the Service Description!

Change-Id: If5e2ae3e1c8e54948f10fb1e16c1e5b776fc61b5