4204 FIX Prevent password completion for all WATO rule where passwords are being specified
Most browsers try to autofill password input fields in forms.
While this is useful in most times it is very nasty in WATO
rules where passwords are being configured. This not only
destroys existing passwords when loading and saving a rule. It
also renders the {{view in plaintext}} button useless.

This has been fixed for all browsers.

Change-Id: Idcc0debe1926e931aca85d1e47e230b8dd55e7bd