4362 fileinfo.groups, logwatch.groups: now group pattern matching is allowed in service description
The service description may contain one or more occurances of %s. If you do this,
then the pattern must be a regular expression and be prefixed with ~. For each %s
in the description, the expression has to contain one "group". A group is a
subexpression enclosed in brackets, for example (.*) or ([a-zA-Z]+) or (...).

Change-Id: I69ec4a51ebd18e766aa116cefcfaae03e4805f9f