4436 FIX check_sql: fixed thresholds and added missing options
Thresholds are now actually used when the options are activated and the number is added to the output if Thresholds are set. Additionally the option "Performance Data" is now actually used. If not activated no performance data will be added, even if it is given by the result of the SQL statement.

Change-Id: Ifb3780f318f3a2bc81ade56007698bbc7fd43f16