4396 FIX Fixed exception in BI views when using site icons
There is a annicient and now hidden feature "site icons". The user can set the option
use_siteicons = True in the multisite.mk and put an icon named like this into the
images/icons directory: icons/site-%s-24.png.

When this is enabled and the sites icon painter is added to a BI view, this leads
to an exception. That the "site" key is not available.

Change-Id: I09f6f678e117dad121f9ec23a05e82f8c07c43bc