4420 FIX Cleaning up signatures of unused packages now
The signatures of old agents (packages that are not needed by any host) were not
deleted from the var/check_mk/agents/*/_PACKAGES directories while the packages
and config files were deleted.

This had the effect that when you change a agent package relevant configuration,
bake the agent, change the value back and bake the agent with the same config,
the agent will already be signed without the need manually sign it again. This
was possible when the former signed config is exactly the same as the new one.

Change-Id: I6233f24fa95e157095a729b0eaa698db46f741df