4422 logwatch.ec: Message forwarding via TCP can now use a spool
When forwaring messages to a remote host via TCP it can happen that
messages are dropped. This may happen e.g. when the destination host
or service is not available. By default the messages to be forwarded
are dropped.

Now you can configure the TCP forwarding to spool the messages until
a deadline or the spool has reached a specific size to prevent
loss of messages.

Change-Id: I22bd16c3f11ee4ab872c634ea7f5b078f8327e20