4452 Multisite Quicksearch: Now able to configure quicksearch filters in the global settings
You can now specifiy which quicksearch filters should be automatically traversed
whenever a text is entered in the quicksearch field. The default behaviour is
<ul>
<li>h: Hostname</li>
<li>al: Hostalias</li>
<li>ad: Hostaddress</li>
<li>s: Service description</li>
</ul

Results for each filter are now shown in separate topics, when there are matches
for multiple filters. You can also enable filters which are currently not used
in the default search
<ul>
<li>tg: Hosttag</li>
<li>hg: Hostgroup</li>
<li>sg: Servicegroup</li>
</ul

Additionally, you can modify the matching behaviour for a specific filter.
When a match is found you can continue the search, abort the search and show all results
or abort the search and show only the results for the current filter.

Change-Id: I7bb8594bba173ae96eabc20e8cb1998439f74183