4385 FIX mk_inventory.vbs: Add new query type for installed software on hosts
We added a new query type to get the information about installed software on hosts.
This new type is the default way, how information is collected.
The old way is a fallback and will be tried, if the new one is not possible.

In error situations, we do not query the installed software and report an empty section.

Change-Id: I691974af4abfd8af97685d7e4e63658dd8c4efe2