4423 FIX Availability of service filtered views was showing services multiple times
When a service is in multiple service groups and a view is filtered by a single service
group and the availability view is grouped by service groups, a single service was
shown multiple times. Once for each service group membership.

The service is now shown only for the filtered service group memberships.

Change-Id: I9ad374abe29d6f7cf1dca24201281ddb99be3e87