4499 FIX Agent Updater: Prevent broken state files caused by incomplete writes
When the agent updater is being executed it writes it's current deployment state
to a state file. When the system is turned of suddenly it could cause a file
corruption of the cmk-update-agent.state file. We prevent this now by writing to
a temporary file first, flush it and then move it over to the state file path.

Change-Id: Ic8075794d25e0fe90be32bda03632525a43b2880