Added MonitoringCore::impl().
This effectively acts as an escape hatch to access the underlying
implementation.  While this is not particularly nice and should die in the
long run, it is very handy in the meantime, reducing the need for parameter
passing quite a bite.  Note that even the C++ standard library has similar
hacks, e.g. std::thread::native_handle():

Change-Id: I47ff2650da2e2e6abac845adf1a7487e60712325