4540 emcvnx_disks: Thresholds for count of failures can be set now
Previously only the state for hard read or write failures could be set. As from
now additionally a counter of failures can be set. The Service will change the
state only if the amount of failure will be equal or higher than set in rule.

On default the Service will be go critical if there are 2 or more failures
detected.

Change-Id: Id71eee97c51c8f75ac29c04eff3573ec66c96bc0