8627 FIX apt: Fix Ubuntu "NOTE: This is only a simulation!" warning being output in check
When running the Check_MK agent as a non-root user, the apt-get command used in the
mk_apt plugin would output a note that would then appear in the update check.
The agent plugin no longer outputs this message. To fix the problem, you will need
to deploy the updated version of the mk_apt plugin.

Thanks to Barnaby Shearer for the fix!

Change-Id: Ic702a3d474da8d3e57c52d6f7b6ca0bd66d1c0ea