4470 FIX Event console: fixed incorrect event counting for timed-out and heartbeat events
The internal event counter of status_num_open_events (and others) was not correctly decreased for

<ul>
<li>events with timeouts, e.g. an event with a lifetime</li>
<li>events which expect something in a given interval</li>
</ul>

Since this event count never decreased, the limits specified in the global settings rule <i>Limit amount of current events</i>,
were reached when the event console ran for a certain amount of time. A restart always fixed this problem, because these counts
are calculated from scratch during startup.

Change-Id: Iab41625af9619217c80751933e196d87b9dacbd2