4435 FIX Fixed pending_flex_downtime column in host/service-related tables
In the NEB plugin, the pending_flex_downtime column has always been an
integer, although the column comment claimed it was a boolean.  The
Livestatus interface of the Check_MK Micro Core was inconsistent and
returned a boolean.  Furthermore, the calculation of the value was not
totally correct.

Now pending_flex_downtime is always returned as an integer, and the column
comment has been fixed to reflect that.

Change-Id: I9ae8a13e651492a3421ad89d814c0d72fb92e690