4625 FIX mk_oracle: Fixed configuration for remote instances in bakery
There has been a mistake causing an incorrect configuration for remote
instances. In the confiugration created by the agent bakery the information for
the piggyhost was missing. With this a connection could not be established,
because of the consequential errors. Also this error has always been present
since the bakery feature for oracle has been introduced.

Change-Id: If5b1b61f031b8564706790d347f0f7ff134c45d3