4671 FIX Metrics: Fixed exceptions when trying to parse perfdata with range warn/crit
The Nagios performance data format allows plugins to specify ranges for warn/crit thresholds.
The numbers need to be separated by a ":" in this case. This is not supported with the Check_MK
graphing system and lead to an exception.

The handling has now been changed to silently ignore those values.

Change-Id: I6d2f3cb0dd6afca6d7c2343f47f221bf3303ab1a