4725 FIX Significant improvement of Check_MK check helper memory usage
The memory usage of the Check_MK check helpers, especially the growth
that occures during runtime, has been reduced significantly.

One example which has been measured in a single site with 30 Check_MK
check helpers configured and a large configuration
(a lot of hosts, services, tags and rulesets):

Before this change it was executing 1000 service checks per second with
a total memory usage of about 15 GB.

After this change it is now performing 1100 service checks per second
with a total memory usage of 3.5(!) GB.

The exact memory reduction highly depends on the Check_MK configuration,
but should be measurable in all medium to large environments.

Change-Id: I42e2a6f9da071b0a5f302bb41d7ae1db973fccc3