4709 FIX Notifications: Missing SERVICECONTACTGROUPNAMES/HOSTCONTACTGROUPNAMES parameters no longer causes a notification not to be sent
This werk reverts part of WK3886. The nagios core did not set the environment parameters mentioned in the topic.
As a result the notifcation condition <i>Match Contact Groups</i> did never match and a notification was not sent.
This has been changed so that this condition check is skipped, whenever this information is missing.

An error scenario should always let the message through.

The issue, that the nagios core does not set this environment info will be fixed in a future werk.

Change-Id: If330de216270299a54f406a913e654526d540381