4758 FIX Fixed site status and master control snapin not displaying states of sites
The master control snapin was hiding sites with communication errors completely.
It is now displaying a box with the occured error message.

The site status snapin was showing the site, but displaying a site as "online"
while it was reported as "dead" by the livestatus connection. This appeared when
an error was found in a query, for example when a queried table does not exist.

Change-Id: I44aa10019a7b78fa381bdb3ebac1719c4ab22576