4892 FIX Cleanup site changes of deleted sites correctly
When a site was replicated via WATO before and is deleted, the file where
the pending changes of that site were stored was left in the master site.

The path to the file is "var/check_mk/wato/replication_changes_[site].mk".
In case you have files for old sites you can savely delete them.

Change-Id: I98eb375e43300fc5f9551987b4fc703c2f96d0d6