Add a configure flag for old librrd versions.
Old (pre-1.6.0) versions of librrd were not thread-safe, so one has to use the
thread-safe variant librrd_th explicitly. This in itself is already tragic
enough, but to make things worse, you can't write a compile-time check to
distinguish these versions. :-/ So we default to the assumption that librrd is
thread-safe (which is the case within OMD) and the user of e.g. a standalone
version of Livestatus on an old system has to use the configure flag below.
Not really nice, but we can't really do much about that.

Change-Id: Icc1132814eaf8bd621aa7a0e5f981b1cd48aa625