4838 FIX netapp_api_volumes: No longer shows service descriptions with UUIDs
The datasource program rule for the netapp agent offers an option to skip
the performance counters for volumes. When this option was set (default off)
the service description changed from a <i>vserver_name.volume_name</i> style
to the <i>uuid</i>. This has been fixed. You can now set the counters "off"
without the risk of a changed service description.

Change-Id: I40c2c67f1f33ebc0766585024c44ed2bf406f6f5