4921 Users can be limited to a list of authorized sites now
It is now possible to configure a list of authorized sites for each user.

This is especially useful when having a huge number of sites configured,
like several hundred, in a distributed setup. In case you have users that
have their hosts on only some of these sites, you can set the new authorized
sites option for these users.

For the users that have authorized sites configured the GUI will only contact
these sites instead of all sites to build up their views. In large environments
this will be a huge performance improvement.

Change-Id: I428928d99714dda7dffa2a668cd63b67830ce610